requirements.txt
Django==5.1.5
psycopg2-binary==2.9.10
djangorestframework==3.15.2
python_dotenv==1.0.1
djangorestframework_simplejwt==5.4.0
django-cors-headers==4.6.0
pytest==8.3.4
pytest-django==4.10.0
